Change code to enable role assumption
